% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_mdlib.R
\name{connect_mdlib}
\alias{connect_mdlib}
\title{connect_mdlib}
\usage{
connect_mdlib(path, query = NULL, token)
}
\arguments{
\item{path}{character: path to send request}

\item{query}{character}

\item{token}{character: access token to be opbtained from \url{http://microdatalibqa.worldbank.org/apidocumentation/}}
}
\value{
object of class mdlib_api
}
\description{
connect_mdlib
}
