% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_dictionary.R
\name{get_data_dictionary}
\alias{get_data_dictionary}
\title{get_data_dictionary}
\usage{
get_data_dictionary(id, token, limit = NULL)
}
\arguments{
\item{id}{numeric: Survey Unique ID}

\item{token}{character: Microdata API authentication token}

\item{limit}{character: maximum number of records per page}
}
\value{
list
}
\description{
get_data_dictionary
}
