% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datafiles_list.R
\name{get_datafiles_list}
\alias{get_datafiles_list}
\title{get_datafiles_list}
\usage{
get_datafiles_list(id, token)
}
\arguments{
\item{id}{numeric: Survey Unique ID}

\item{token}{character}
}
\value{
list
}
\description{
get_datafiles_list
}
