% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_doc_metadata.R
\name{get_doc_metadata}
\alias{get_doc_metadata}
\title{get_doc_metadata}
\usage{
get_doc_metadata(id, datafile_id, token)
}
\arguments{
\item{id}{numeric: Survey Unique ID}

\item{datafile_id}{character: Data file Unique ID}

\item{token}{character}
}
\value{
list
}
\description{
get_doc_metadata
}
