% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_public_survey_list.R
\name{get_public_survey_list}
\alias{get_public_survey_list}
\title{get_public_survey_list}
\usage{
get_public_survey_list(token)
}
\arguments{
\item{token}{character}
}
\value{
list
}
\description{
get_public_survey_list
}
