% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_metadata.R
\name{get_survey_metadata}
\alias{get_survey_metadata}
\title{get_survey_metadata}
\usage{
get_survey_metadata(id, token)
}
\arguments{
\item{id}{numeric: Survey Unique ID}

\item{token}{character}
}
\value{
list
}
\description{
get_survey_metadata
}
