% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_metadata.R
\name{get_variable_metadata}
\alias{get_variable_metadata}
\title{get_variable_metadata}
\usage{
get_variable_metadata(id, variable_id, token)
}
\arguments{
\item{id}{numeric: Survey Unique ID}

\item{variable_id}{character: Variable Unique ID}

\item{token}{character}
}
\value{
list
}
\description{
get_variable_metadata
}
