% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables_by_study.R
\name{get_variables_by_study}
\alias{get_variables_by_study}
\title{get_variables_by_study}
\usage{
get_variables_by_study(id, token)
}
\arguments{
\item{id}{numeric: Survey Unique ID}

\item{token}{character}
}
\value{
list
}
\description{
get_variables_by_study
}
