% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables_list.R
\name{get_variables_list}
\alias{get_variables_list}
\title{get_variables_list}
\usage{
get_variables_list(id, datafile_id, token)
}
\arguments{
\item{id}{numeric: Survey Unique ID}

\item{datafile_id}{character: Data file Unique ID}

\item{token}{character}
}
\value{
list
}
\description{
get_variables_list
}
