% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_mdlib.R
\name{connect_mdlib}
\alias{connect_mdlib}
\title{connect_mdlib}
\usage{
connect_mdlib(
  path,
  query = NULL,
  token,
  root = "https://microdatalib.worldbank.org"
)
}
\arguments{
\item{path}{character: path to send request}

\item{query}{character}

\item{token}{character: access token to be obtained from \url{https://microdatalib.worldbank.org/index.php/auth/api_keys}}

\item{root}{character: the root url that hosts the API}
}
\value{
object of class mdlib_api
}
\description{
connect_mdlib
}
